feat: Add upcoming birthdays functionality to AddressBook
The code changes introduce a new method `get_upcoming_birthdays` to the `AddressBook` class. This method retrieves a list of upcoming birthdays within a specified number of days. The method takes an optional parameter `days` to specify the number of days to consider for upcoming birthdays, with a default value of 7. The method returns a list of tuples containing the name and date of upcoming birthdays.

This commit message follows the established convention of using a prefix to indicate the type of change, followed by a concise and descriptive message.